textures working again
Removed a "static" keyword from the texture array declaration when I
tried to transition to the Star object. When I tried to revert, forgot
about static, so the array was being reset to 0 every update, but the
stars were only being draw once when the timing ran out.

Made the loop that populates the texture array run every update so that
static doesn't matter. I'm planning on having it not be static anyway,
and just having the Star object be static.